review: move stackless into utils/error.js + add changeset
- merge stackless into the existing utils/error.js (per review) and import it
  directly from there; delete the standalone utils/stackless.js
- harden the regression test to follow @sveltejs/kit/* self-imports, so a Vite
  pull reaching internal/env *through* @sveltejs/kit/internal can't slip past
- add the patch changeset (explicit env vars is now released)
