+++ .changeset/no-vite-in-env-validator.md
@@ -1,0 +1,5 @@
+---
+'@sveltejs/kit': patch
+---
+
+fix: avoid importing Vite while validating explicit environment variables
