+++ packages/kit/src/utils/error.js
@@ -1,6 +1,18 @@
 import { HttpError, SvelteKitError } from '@sveltejs/kit/internal';
 
 /**
+ * For times when you need to throw an error, but without
+ * displaying a useless stack trace (since the developer
+ * can't do anything useful with it)
+ * @param {string} message
+ */
+export function stackless(message) {
+	const error = new Error(message);
+	error.stack = '';
+	return error;
+}
+
+/**
  * @param {unknown} err
  * @return {Error}
  */
